// src/admin/modules/config/submodules/repairCompany/types/repairCompany.ts
export interface RepairCompany {
    id: number;
    name?: string;
    billAddress?: string;
    billCity?: string;
    billState?: string;
    billPin?: string;
    billPhone?: string;
    billEmail?: string;
    billAccountNo?: string;
    shipAddress?: string;
    shipCity?: string;
    shipState?: string;
    shipPin?: string;
    shipEmail?: string;
    shipPhone?: string;
    shipAccountNo?: string;
  }